+++ comp/forwarder/defaultforwarder/params.go
@@ -10,6 +10,7 @@ import (
 	"github.com/DataDog/datadog-agent/comp/core/log"
 	pkgconfig "github.com/DataDog/datadog-agent/pkg/config"
 	"github.com/DataDog/datadog-agent/pkg/config/resolver"
+	"github.com/DataDog/datadog-agent/pkg/config/utils"
 )
 
 type Params struct {
@@ -30,7 +31,7 @@ func NewParamsWithResolvers(config config.Component, log log.Component) Params {
 
 func getMultipleEndpoints(_ config.Component, log log.Component) map[string][]string {
 	// Inject the config to make sure we can call GetMultipleEndpoints.
-	keysPerDomain, err := pkgconfig.GetMultipleEndpoints()
+	keysPerDomain, err := utils.GetMultipleEndpoints(pkgconfig.Datadog)
 	if err != nil {
 		log.Error("Misconfiguration of agent endpoints: ", err)
 	}
