+++ pkg/config/config_test.go
@@ -6,13 +6,10 @@
 package config
 
 import (
-	"bytes"
 	"fmt"
 	"io/ioutil"
-	"log"
 	"os"
 	"path/filepath"
-	"strings"
 	"testing"
 	"time"
 
@@ -21,22 +18,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
-
-func setupConf() Config {
-	conf := NewConfig("datadog", "DD", strings.NewReplacer(".", "_"))
-	InitConfig(conf)
-	return conf
-}
-
-func setupConfFromYAML(yamlConfig string) Config {
-	conf := setupConf()
-	conf.SetConfigType("yaml")
-	e := conf.ReadConfig(bytes.NewBuffer([]byte(yamlConfig)))
-	if e != nil {
-		log.Println(e)
-	}
-	return conf
-}
 
 func unsetEnvForTest(t *testing.T, env string) {
 	oldValue, ok := os.LookupEnv(env)
@@ -52,7 +33,7 @@ func unsetEnvForTest(t *testing.T, env string) {
 }
 
 func TestDefaults(t *testing.T) {
-	config := setupConf()
+	config := SetupConf()
 
 	// Testing viper's handling of defaults
 	assert.False(t, config.IsSet("site"))
@@ -69,52 +50,11 @@ func TestDefaults(t *testing.T) {
 	}, config.GetStringMap("process_config.process_discovery"))
 }
 
-func TestDefaultSite(t *testing.T) {
-	datadogYaml := `
-api_key: fakeapikey
-`
-	testConfig := setupConfFromYAML(datadogYaml)
-
-	multipleEndpoints, err := getMultipleEndpointsWithConfig(testConfig)
-	externalAgentURL := GetMainEndpointWithConfig(testConfig, "https://external-agent.", "external_config.external_agent_dd_url")
-
-	expectedMultipleEndpoints := map[string][]string{
-		"https://app.datadoghq.com": {
-			"fakeapikey",
-		},
-	}
-
-	assert.NoError(t, err)
-	assert.EqualValues(t, expectedMultipleEndpoints, multipleEndpoints)
-	assert.Equal(t, "https://external-agent.datadoghq.com", externalAgentURL)
-}
-
-func TestSite(t *testing.T) {
-	datadogYaml := `
-site: datadoghq.eu
-api_key: fakeapikey
-`
-	testConfig := setupConfFromYAML(datadogYaml)
-
-	multipleEndpoints, err := getMultipleEndpointsWithConfig(testConfig)
-	externalAgentURL := GetMainEndpointWithConfig(testConfig, "https://external-agent.", "external_config.external_agent_dd_url")
-
-	expectedMultipleEndpoints := map[string][]string{
-		"https://app.datadoghq.eu": {
-			"fakeapikey",
-		},
-	}
-
-	assert.NoError(t, err)
-	assert.EqualValues(t, expectedMultipleEndpoints, multipleEndpoints)
-	assert.Equal(t, "https://external-agent.datadoghq.eu", externalAgentURL)
-}
-
 func TestUnexpectedUnicode(t *testing.T) {
 	keyYaml := "api_\u202akey: fakeapikey\n"
 	valueYaml := "api_key: fa\u202akeapikey\n"
 
-	testConfig := setupConfFromYAML(keyYaml)
+	testConfig := SetupConfFromYAML(keyYaml)
 
 	warnings := findUnexpectedUnicode(testConfig)
 	require.Len(t, warnings, 1)
@@ -122,7 +62,7 @@ func TestUnexpectedUnicode(t *testing.T) {
 	assert.Contains(t, warnings[0], "Configuration key string")
 	assert.Contains(t, warnings[0], "U+202A")
 
-	testConfig = setupConfFromYAML(valueYaml)
+	testConfig = SetupConfFromYAML(valueYaml)
 
 	warnings = findUnexpectedUnicode(testConfig)
 
@@ -133,7 +73,7 @@ func TestUnexpectedUnicode(t *testing.T) {
 
 func TestUnexpectedNestedUnicode(t *testing.T) {
 	yaml := "runtime_security_config:\n  activity_dump:\n    remote_storage:\n      endpoints:\n        logs_dd_url: \"http://\u202adatadawg.com\""
-	testConfig := setupConfFromYAML(yaml)
+	testConfig := SetupConfFromYAML(yaml)
 
 	warnings := findUnexpectedUnicode(testConfig)
 	require.Len(t, warnings, 1)
@@ -165,7 +105,7 @@ func TestUnexpectedWhitespace(t *testing.T) {
 		},
 	}
 	for _, tc := range tests {
-		testConfig := setupConfFromYAML(tc.yaml)
+		testConfig := SetupConfFromYAML(tc.yaml)
 		warnings := findUnexpectedUnicode(testConfig)
 		require.Len(t, warnings, 1)
 
@@ -178,14 +118,14 @@ func TestUnknownKeysWarning(t *testing.T) {
 	yamlBase := `
 site: datadoghq.eu
 `
-	confBase := setupConfFromYAML(yamlBase)
+	confBase := SetupConfFromYAML(yamlBase)
 	assert.Len(t, findUnknownKeys(confBase), 0)
 
 	yamlWithUnknownKeys := `
 site: datadoghq.eu
 unknown_key.unknown_subkey: true
 `
-	confWithUnknownKeys := setupConfFromYAML(yamlWithUnknownKeys)
+	confWithUnknownKeys := SetupConfFromYAML(yamlWithUnknownKeys)
 	assert.Len(t, findUnknownKeys(confWithUnknownKeys), 1)
 
 	confWithUnknownKeys.SetKnown("unknown_key.*")
@@ -214,362 +154,23 @@ func TestUnknownVarsWarning(t *testing.T) {
 	t.Run("DD_SYSTEM_PROBE_EXTRA", test("DD_SYSTEM_PROBE_EXTRA", false, []string{"DD_SYSTEM_PROBE_EXTRA"}))
 }
 
-func TestSiteEnvVar(t *testing.T) {
-	t.Setenv("DD_API_KEY", "fakeapikey")
-	t.Setenv("DD_SITE", "datadoghq.eu")
-	testConfig := setupConfFromYAML("")
-
-	multipleEndpoints, err := getMultipleEndpointsWithConfig(testConfig)
-	externalAgentURL := GetMainEndpointWithConfig(testConfig, "https://external-agent.", "external_config.external_agent_dd_url")
-
-	expectedMultipleEndpoints := map[string][]string{
-		"https://app.datadoghq.eu": {
-			"fakeapikey",
-		},
-	}
-
-	assert.NoError(t, err)
-	assert.EqualValues(t, expectedMultipleEndpoints, multipleEndpoints)
-	assert.Equal(t, "https://external-agent.datadoghq.eu", externalAgentURL)
-}
-
 func TestDefaultTraceManagedServicesEnvVarValue(t *testing.T) {
-	testConfig := setupConfFromYAML("")
+	testConfig := SetupConfFromYAML("")
 	assert.Equal(t, true, testConfig.Get("serverless.trace_managed_services"))
 }
 
 func TestExplicitFalseTraceManagedServicesEnvVar(t *testing.T) {
 	t.Setenv("DD_TRACE_MANAGED_SERVICES", "false")
-	testConfig := setupConfFromYAML("")
+	testConfig := SetupConfFromYAML("")
 	assert.Equal(t, false, testConfig.Get("serverless.trace_managed_services"))
 }
 
 func TestDDHostnameFileEnvVar(t *testing.T) {
 	t.Setenv("DD_API_KEY", "fakeapikey")
 	t.Setenv("DD_HOSTNAME_FILE", "somefile")
-	testConfig := setupConfFromYAML("")
+	testConfig := SetupConfFromYAML("")
 
 	assert.Equal(t, "somefile", testConfig.Get("hostname_file"))
-}
-
-func TestDDURLEnvVar(t *testing.T) {
-	t.Setenv("DD_API_KEY", "fakeapikey")
-	t.Setenv("DD_URL", "https://app.datadoghq.eu")
-	t.Setenv("DD_EXTERNAL_CONFIG_EXTERNAL_AGENT_DD_URL", "https://custom.external-agent.datadoghq.com")
-	testConfig := setupConfFromYAML("")
-	testConfig.BindEnv("external_config.external_agent_dd_url")
-
-	multipleEndpoints, err := getMultipleEndpointsWithConfig(testConfig)
-	externalAgentURL := GetMainEndpointWithConfig(testConfig, "https://external-agent.", "external_config.external_agent_dd_url")
-
-	expectedMultipleEndpoints := map[string][]string{
-		"https://app.datadoghq.eu": {
-			"fakeapikey",
-		},
-	}
-
-	assert.NoError(t, err)
-	assert.EqualValues(t, expectedMultipleEndpoints, multipleEndpoints)
-	assert.Equal(t, "https://custom.external-agent.datadoghq.com", externalAgentURL)
-}
-
-func TestDDDDURLEnvVar(t *testing.T) {
-	t.Setenv("DD_API_KEY", "fakeapikey")
-	t.Setenv("DD_DD_URL", "https://app.datadoghq.eu")
-	t.Setenv("DD_EXTERNAL_CONFIG_EXTERNAL_AGENT_DD_URL", "https://custom.external-agent.datadoghq.com")
-	testConfig := setupConfFromYAML("")
-	testConfig.BindEnv("external_config.external_agent_dd_url")
-
-	multipleEndpoints, err := getMultipleEndpointsWithConfig(testConfig)
-	externalAgentURL := GetMainEndpointWithConfig(testConfig, "https://external-agent.", "external_config.external_agent_dd_url")
-
-	expectedMultipleEndpoints := map[string][]string{
-		"https://app.datadoghq.eu": {
-			"fakeapikey",
-		},
-	}
-
-	assert.NoError(t, err)
-	assert.EqualValues(t, expectedMultipleEndpoints, multipleEndpoints)
-	assert.Equal(t, "https://custom.external-agent.datadoghq.com", externalAgentURL)
-}
-
-func TestDDURLAndDDDDURLEnvVar(t *testing.T) {
-	t.Setenv("DD_API_KEY", "fakeapikey")
-
-	// If DD_DD_URL and DD_URL are set, the value of DD_DD_URL is used
-	t.Setenv("DD_DD_URL", "https://app.datadoghq.dd_dd_url.eu")
-	t.Setenv("DD_URL", "https://app.datadoghq.dd_url.eu")
-
-	t.Setenv("DD_EXTERNAL_CONFIG_EXTERNAL_AGENT_DD_URL", "https://custom.external-agent.datadoghq.com")
-	testConfig := setupConfFromYAML("")
-	testConfig.BindEnv("external_config.external_agent_dd_url")
-
-	multipleEndpoints, err := getMultipleEndpointsWithConfig(testConfig)
-	externalAgentURL := GetMainEndpointWithConfig(testConfig, "https://external-agent.", "external_config.external_agent_dd_url")
-
-	expectedMultipleEndpoints := map[string][]string{
-		"https://app.datadoghq.dd_dd_url.eu": {
-			"fakeapikey",
-		},
-	}
-
-	assert.NoError(t, err)
-	assert.EqualValues(t, expectedMultipleEndpoints, multipleEndpoints)
-	assert.Equal(t, "https://custom.external-agent.datadoghq.com", externalAgentURL)
-}
-
-func TestDDURLOverridesSite(t *testing.T) {
-	datadogYaml := `
-site: datadoghq.eu
-dd_url: "https://app.datadoghq.com"
-api_key: fakeapikey
-
-external_config:
-  external_agent_dd_url: "https://external-agent.datadoghq.com"
-`
-	testConfig := setupConfFromYAML(datadogYaml)
-
-	multipleEndpoints, err := getMultipleEndpointsWithConfig(testConfig)
-	externalAgentURL := GetMainEndpointWithConfig(testConfig, "https://external-agent.", "external_config.external_agent_dd_url")
-
-	expectedMultipleEndpoints := map[string][]string{
-		"https://app.datadoghq.com": {
-			"fakeapikey",
-		},
-	}
-
-	assert.NoError(t, err)
-	assert.EqualValues(t, expectedMultipleEndpoints, multipleEndpoints)
-	assert.Equal(t, "https://external-agent.datadoghq.com", externalAgentURL)
-}
-
-func TestDDURLNoSite(t *testing.T) {
-	datadogYaml := `
-dd_url: "https://app.datadoghq.eu"
-api_key: fakeapikey
-
-external_config:
-  external_agent_dd_url: "https://custom.external-agent.datadoghq.eu"
-`
-	testConfig := setupConfFromYAML(datadogYaml)
-
-	multipleEndpoints, err := getMultipleEndpointsWithConfig(testConfig)
-	externalAgentURL := GetMainEndpointWithConfig(testConfig, "https://external-agent.", "external_config.external_agent_dd_url")
-
-	expectedMultipleEndpoints := map[string][]string{
-		"https://app.datadoghq.eu": {
-			"fakeapikey",
-		},
-	}
-
-	assert.NoError(t, err)
-	assert.EqualValues(t, expectedMultipleEndpoints, multipleEndpoints)
-	assert.Equal(t, "https://custom.external-agent.datadoghq.eu", externalAgentURL)
-}
-
-func TestGetMultipleEndpointsDefault(t *testing.T) {
-	datadogYaml := `
-api_key: fakeapikey
-
-additional_endpoints:
-  "https://app.datadoghq.com":
-  - fakeapikey2
-  - fakeapikey3
-  "https://foo.datadoghq.com":
-  - someapikey
-`
-
-	testConfig := setupConfFromYAML(datadogYaml)
-
-	multipleEndpoints, err := getMultipleEndpointsWithConfig(testConfig)
-
-	expectedMultipleEndpoints := map[string][]string{
-		"https://foo.datadoghq.com": {
-			"someapikey",
-		},
-		"https://app.datadoghq.com": {
-			"fakeapikey",
-			"fakeapikey2",
-			"fakeapikey3",
-		},
-	}
-
-	assert.NoError(t, err)
-	assert.EqualValues(t, expectedMultipleEndpoints, multipleEndpoints)
-}
-
-func TestGetMultipleEndpointsEnvVar(t *testing.T) {
-	t.Setenv("DD_API_KEY", "fakeapikey")
-	t.Setenv("DD_ADDITIONAL_ENDPOINTS", "{\"https://foo.datadoghq.com\": [\"someapikey\"]}")
-
-	testConfig := setupConf()
-
-	multipleEndpoints, err := getMultipleEndpointsWithConfig(testConfig)
-
-	expectedMultipleEndpoints := map[string][]string{
-		"https://foo.datadoghq.com": {
-			"someapikey",
-		},
-		"https://app.datadoghq.com": {
-			"fakeapikey",
-		},
-	}
-
-	assert.NoError(t, err)
-	assert.EqualValues(t, expectedMultipleEndpoints, multipleEndpoints)
-}
-
-func TestGetMultipleEndpointsSite(t *testing.T) {
-	datadogYaml := `
-site: datadoghq.eu
-api_key: fakeapikey
-
-additional_endpoints:
-  "https://app.datadoghq.com":
-  - fakeapikey2
-  - fakeapikey3
-  "https://foo.datadoghq.com":
-  - someapikey
-`
-
-	testConfig := setupConfFromYAML(datadogYaml)
-
-	multipleEndpoints, err := getMultipleEndpointsWithConfig(testConfig)
-
-	expectedMultipleEndpoints := map[string][]string{
-		"https://app.datadoghq.eu": {
-			"fakeapikey",
-		},
-		"https://foo.datadoghq.com": {
-			"someapikey",
-		},
-		"https://app.datadoghq.com": {
-			"fakeapikey2",
-			"fakeapikey3",
-		},
-	}
-
-	assert.NoError(t, err)
-	assert.EqualValues(t, expectedMultipleEndpoints, multipleEndpoints)
-}
-
-func TestGetMultipleEndpointsDDURL(t *testing.T) {
-	datadogYaml := `
-dd_url: "https://app.datadoghq.com"
-api_key: fakeapikey
-
-additional_endpoints:
-  "https://app.datadoghq.com":
-  - fakeapikey2
-  - fakeapikey3
-  "https://foo.datadoghq.com":
-  - someapikey
-`
-
-	testConfig := setupConfFromYAML(datadogYaml)
-
-	multipleEndpoints, err := getMultipleEndpointsWithConfig(testConfig)
-
-	expectedMultipleEndpoints := map[string][]string{
-		"https://foo.datadoghq.com": {
-			"someapikey",
-		},
-		"https://app.datadoghq.com": {
-			"fakeapikey",
-			"fakeapikey2",
-			"fakeapikey3",
-		},
-	}
-
-	assert.NoError(t, err)
-	assert.EqualValues(t, expectedMultipleEndpoints, multipleEndpoints)
-}
-
-func TestGetMultipleEndpointsWithNoAdditionalEndpoints(t *testing.T) {
-	datadogYaml := `
-dd_url: "https://app.datadoghq.com"
-api_key: fakeapikey
-`
-
-	testConfig := setupConfFromYAML(datadogYaml)
-
-	multipleEndpoints, err := getMultipleEndpointsWithConfig(testConfig)
-
-	expectedMultipleEndpoints := map[string][]string{
-		"https://app.datadoghq.com": {
-			"fakeapikey",
-		},
-	}
-
-	assert.NoError(t, err)
-	assert.EqualValues(t, expectedMultipleEndpoints, multipleEndpoints)
-}
-
-func TestGetMultipleEndpointseIgnoresDomainWithoutApiKey(t *testing.T) {
-	datadogYaml := `
-dd_url: "https://app.datadoghq.com"
-api_key: fakeapikey
-
-additional_endpoints:
-  "https://app.datadoghq.com":
-  - fakeapikey2
-  "https://foo.datadoghq.com":
-  - someapikey
-  "https://bar.datadoghq.com":
-  - ""
-`
-
-	testConfig := setupConfFromYAML(datadogYaml)
-
-	multipleEndpoints, err := getMultipleEndpointsWithConfig(testConfig)
-
-	expectedMultipleEndpoints := map[string][]string{
-		"https://app.datadoghq.com": {
-			"fakeapikey",
-			"fakeapikey2",
-		},
-		"https://foo.datadoghq.com": {
-			"someapikey",
-		},
-	}
-
-	assert.NoError(t, err)
-	assert.EqualValues(t, expectedMultipleEndpoints, multipleEndpoints)
-}
-
-func TestGetMultipleEndpointsApiKeyDeduping(t *testing.T) {
-	datadogYaml := `
-dd_url: "https://app.datadoghq.com"
-api_key: fakeapikey
-
-additional_endpoints:
-  "https://app.datadoghq.com":
-  - fakeapikey2
-  - fakeapikey
-  "https://foo.datadoghq.com":
-  - someapikey
-  - someotherapikey
-  - someapikey
-`
-
-	testConfig := setupConfFromYAML(datadogYaml)
-
-	multipleEndpoints, err := getMultipleEndpointsWithConfig(testConfig)
-
-	expectedMultipleEndpoints := map[string][]string{
-		"https://app.datadoghq.com": {
-			"fakeapikey",
-			"fakeapikey2",
-		},
-		"https://foo.datadoghq.com": {
-			"someapikey",
-			"someotherapikey",
-		},
-	}
-
-	assert.NoError(t, err)
-	assert.EqualValues(t, expectedMultipleEndpoints, multipleEndpoints)
 }
 
 func TestAddAgentVersionToDomain(t *testing.T) {
@@ -683,7 +284,7 @@ func TestIsCloudProviderEnabled(t *testing.T) {
 }
 
 func TestEnvNestedConfig(t *testing.T) {
-	config := setupConf()
+	config := SetupConf()
 	config.BindEnv("foo.bar.nested")
 	t.Setenv("DD_FOO_BAR_NESTED", "baz")
 
@@ -889,7 +490,7 @@ func TestProxy(t *testing.T) {
 			// CircleCI sets NO_PROXY, so unset it for this test
 			unsetEnvForTest(t, "NO_PROXY")
 
-			config := setupConf()
+			config := SetupConf()
 			config.Set("use_proxy_for_cloud_metadata", c.proxyForCloudMetadata)
 
 			// Viper.MergeConfigOverride, which is used when secrets is enabled, will silently fail if a
@@ -912,7 +513,7 @@ func TestProxy(t *testing.T) {
 }
 
 func TestSanitizeAPIKeyConfig(t *testing.T) {
-	config := setupConf()
+	config := SetupConf()
 
 	config.Set("api_key", "foo")
 	SanitizeAPIKeyConfig(config, "api_key")
@@ -931,27 +532,8 @@ func TestSanitizeAPIKeyConfig(t *testing.T) {
 	assert.Equal(t, "foo", config.GetString("api_key"))
 }
 
-// TestSecretBackendWithMultipleEndpoints tests an edge case of `viper.AllSettings()` when a config
-// key includes the key delimiter. Affects the config package when both secrets and multiple
-// endpoints are configured.
-// Refer to https://github.com/DataDog/viper/pull/2 for more details.
-func TestSecretBackendWithMultipleEndpoints(t *testing.T) {
-	conf := setupConf()
-	conf.SetConfigFile("./tests/datadog_secrets.yaml")
-	// load the configuration
-	_, err := LoadDatadogCustom(conf, "datadog_secrets.yaml", true)
-	assert.NoError(t, err)
-
-	expectedKeysPerDomain := map[string][]string{
-		"https://app.datadoghq.com": {"someapikey", "someotherapikey"},
-	}
-	keysPerDomain, err := getMultipleEndpointsWithConfig(conf)
-	assert.NoError(t, err)
-	assert.Equal(t, expectedKeysPerDomain, keysPerDomain)
-}
-
 func TestNumWorkers(t *testing.T) {
-	config := setupConf()
+	config := SetupConf()
 
 	config.Set("python_version", "2")
 	config.Set("tracemalloc_debug", true)
@@ -992,7 +574,7 @@ external_config:
 		"api_key": "overrided",
 	})
 
-	config := setupConfFromYAML(datadogYaml)
+	config := SetupConfFromYAML(datadogYaml)
 	applyOverrideFuncs(config)
 
 	assert.Equal(config.GetString("api_key"), "overrided", "the api key should have been overrided")
@@ -1032,7 +614,7 @@ dogstatsd_mapper_profiles:
         tags:
           foo: "$1"
 `
-	testConfig := setupConfFromYAML(datadogYaml)
+	testConfig := SetupConfFromYAML(datadogYaml)
 
 	profiles, err := getDogstatsdMappingProfilesConfig(testConfig)
 
@@ -1075,7 +657,7 @@ func TestDogstatsdMappingProfilesEmpty(t *testing.T) {
 	datadogYaml := `
 dogstatsd_mapper_profiles:
 `
-	testConfig := setupConfFromYAML(datadogYaml)
+	testConfig := SetupConfFromYAML(datadogYaml)
 
 	profiles, err := getDogstatsdMappingProfilesConfig(testConfig)
 
@@ -1090,7 +672,7 @@ func TestDogstatsdMappingProfilesError(t *testing.T) {
 dogstatsd_mapper_profiles:
   - abc
 `
-	testConfig := setupConfFromYAML(datadogYaml)
+	testConfig := SetupConfFromYAML(datadogYaml)
 	profiles, err := getDogstatsdMappingProfilesConfig(testConfig)
 
 	expectedErrorMsg := "Could not parse dogstatsd_mapper_profiles"
@@ -1115,7 +697,7 @@ func TestDogstatsdMappingProfilesEnv(t *testing.T) {
 }
 
 func TestGetValidHostAliasesWithConfig(t *testing.T) {
-	config := setupConfFromYAML(`host_aliases: ["foo", "-bar"]`)
+	config := SetupConfFromYAML(`host_aliases: ["foo", "-bar"]`)
 	assert.EqualValues(t, getValidHostAliasesWithConfig(config), []string{"foo"})
 }
 
@@ -1123,14 +705,14 @@ func TestNetworkDevicesNamespace(t *testing.T) {
 	datadogYaml := `
 network_devices:
 `
-	config := setupConfFromYAML(datadogYaml)
+	config := SetupConfFromYAML(datadogYaml)
 	assert.Equal(t, "default", config.GetString("network_devices.namespace"))
 
 	datadogYaml = `
 network_devices:
   namespace: dev
 `
-	config = setupConfFromYAML(datadogYaml)
+	config = SetupConfFromYAML(datadogYaml)
 	assert.Equal(t, "dev", config.GetString("network_devices.namespace"))
 }
 
@@ -1154,7 +736,7 @@ logs_config:
   docker_path_override: "/custom/path"
 `
 
-	config := setupConfFromYAML(datadogYaml)
+	config := SetupConfFromYAML(datadogYaml)
 	err := checkConflictingOptions(config)
 
 	assert.NotNil(t, err)
@@ -1205,7 +787,7 @@ proxy:
 `
 	expectedURL := "somehost:1234"
 	expectedHTTPURL := "https://" + expectedURL
-	testConfig := setupConfFromYAML(datadogYaml)
+	testConfig := SetupConfFromYAML(datadogYaml)
 	LoadProxyFromEnv(testConfig)
 	err := setupFipsEndpoints(testConfig)
 	require.NoError(t, err)
@@ -1227,7 +809,7 @@ fips:
 
 	expectedURL = "localhost:50"
 	expectedHTTPURL = "http://" + expectedURL
-	testConfig = setupConfFromYAML(datadogYamlFips)
+	testConfig = SetupConfFromYAML(datadogYamlFips)
 	LoadProxyFromEnv(testConfig)
 	err = setupFipsEndpoints(testConfig)
 	require.NoError(t, err)
@@ -1249,7 +831,7 @@ fips:
 `
 
 	expectedHTTPURL = "https://" + expectedURL
-	testConfig = setupConfFromYAML(datadogYamlFips)
+	testConfig = SetupConfFromYAML(datadogYamlFips)
 	testConfig.Set("skip_ssl_validation", false) // should be overridden by fips.tls_verify
 	LoadProxyFromEnv(testConfig)
 	err = setupFipsEndpoints(testConfig)
@@ -1312,7 +894,7 @@ fips:
   port_range_start: 5000
 `
 
-	testConfig := setupConfFromYAML(datadogYaml)
+	testConfig := SetupConfFromYAML(datadogYaml)
 	err := setupFipsEndpoints(testConfig)
 	require.Error(t, err)
 }
@@ -1322,7 +904,7 @@ func TestEnablePeerServiceStatsAggregationYAML(t *testing.T) {
 apm_config:
   peer_service_aggregation: true
 `
-	testConfig := setupConfFromYAML(datadogYaml)
+	testConfig := SetupConfFromYAML(datadogYaml)
 	err := setupFipsEndpoints(testConfig)
 	require.NoError(t, err)
 	require.True(t, testConfig.GetBool("apm_config.peer_service_aggregation"))
@@ -1331,7 +913,7 @@ apm_config:
 apm_config:
   peer_service_aggregation: false
 `
-	testConfig = setupConfFromYAML(datadogYaml)
+	testConfig = SetupConfFromYAML(datadogYaml)
 	err = setupFipsEndpoints(testConfig)
 	require.NoError(t, err)
 	require.False(t, testConfig.GetBool("apm_config.peer_service_aggregation"))
@@ -1339,10 +921,10 @@ apm_config:
 
 func TestEnablePeerServiceStatsAggregationEnv(t *testing.T) {
 	t.Setenv("DD_APM_PEER_SERVICE_AGGREGATION", "true")
-	testConfig := setupConfFromYAML("")
+	testConfig := SetupConfFromYAML("")
 	require.True(t, testConfig.GetBool("apm_config.peer_service_aggregation"))
 	t.Setenv("DD_APM_PEER_SERVICE_AGGREGATION", "false")
-	testConfig = setupConfFromYAML("")
+	testConfig = SetupConfFromYAML("")
 	require.False(t, testConfig.GetBool("apm_config.peer_service_aggregation"))
 }
 
@@ -1351,7 +933,7 @@ func TestEnableStatsComputationBySpanKindYAML(t *testing.T) {
 apm_config:
   compute_stats_by_span_kind: false
 `
-	testConfig := setupConfFromYAML(datadogYaml)
+	testConfig := SetupConfFromYAML(datadogYaml)
 	err := setupFipsEndpoints(testConfig)
 	require.NoError(t, err)
 	require.False(t, testConfig.GetBool("apm_config.compute_stats_by_span_kind"))
@@ -1360,7 +942,7 @@ apm_config:
 apm_config:
   compute_stats_by_span_kind: true
 `
-	testConfig = setupConfFromYAML(datadogYaml)
+	testConfig = SetupConfFromYAML(datadogYaml)
 	err = setupFipsEndpoints(testConfig)
 	require.NoError(t, err)
 	require.True(t, testConfig.GetBool("apm_config.compute_stats_by_span_kind"))
@@ -1369,9 +951,9 @@ apm_config:
 
 func TestComputeStatsBySpanKindEnv(t *testing.T) {
 	t.Setenv("DD_APM_COMPUTE_STATS_BY_SPAN_KIND", "false")
-	testConfig := setupConfFromYAML("")
+	testConfig := SetupConfFromYAML("")
 	require.False(t, testConfig.GetBool("apm_config.compute_stats_by_span_kind"))
 	t.Setenv("DD_APM_COMPUTE_STATS_BY_SPAN_KIND", "true")
-	testConfig = setupConfFromYAML("")
+	testConfig = SetupConfFromYAML("")
 	require.True(t, testConfig.GetBool("apm_config.compute_stats_by_span_kind"))
 }
