+++ pkg/config/process_test.go
@@ -16,7 +16,7 @@ import (
 
 // TestProcessDefaults tests to ensure that the config has set process settings correctly
 func TestProcessDefaultConfig(t *testing.T) {
-	cfg := setupConf()
+	cfg := SetupConf()
 
 	for _, tc := range []struct {
 		key          string
@@ -135,7 +135,7 @@ func TestProcessDefaultConfig(t *testing.T) {
 
 // TestPrefixes tests that for every corresponding `DD_PROCESS_CONFIG` prefix, there is a `DD_PROCESS_AGENT` prefix as well.
 func TestProcessAgentPrefixes(t *testing.T) {
-	envVarSlice := setupConf().GetEnvVars()
+	envVarSlice := SetupConf().GetEnvVars()
 	envVars := make(map[string]struct{}, len(envVarSlice))
 	for _, envVar := range envVarSlice {
 		envVars[envVar] = struct{}{}
@@ -156,7 +156,7 @@ func TestProcessAgentPrefixes(t *testing.T) {
 
 // TestPrefixes tests that for every corresponding `DD_PROCESS_AGENT` prefix, there is a `DD_PROCESS_CONFIG` prefix as well.
 func TestProcessConfigPrefixes(t *testing.T) {
-	envVarSlice := setupConf().GetEnvVars()
+	envVarSlice := SetupConf().GetEnvVars()
 	envVars := make(map[string]struct{}, len(envVarSlice))
 	for _, envVar := range envVarSlice {
 		envVars[envVar] = struct{}{}
@@ -178,7 +178,7 @@ func TestProcessConfigPrefixes(t *testing.T) {
 }
 
 func TestEnvVarOverride(t *testing.T) {
-	cfg := setupConf()
+	cfg := SetupConf()
 
 	for _, tc := range []struct {
 		key, env, value string
@@ -469,7 +469,7 @@ func readCfgWithType(cfg Config, key, expType string) interface{} {
 }
 
 func TestEnvVarCustomSensitiveWords(t *testing.T) {
-	cfg := setupConf()
+	cfg := SetupConf()
 	expectedPrefixes := []string{"DD_", "DD_PROCESS_CONFIG_", "DD_PROCESS_AGENT_"}
 
 	for i, tc := range []struct {
@@ -501,7 +501,7 @@ func TestEnvVarCustomSensitiveWords(t *testing.T) {
 }
 
 func TestProcBindEnvAndSetDefault(t *testing.T) {
-	cfg := setupConf()
+	cfg := SetupConf()
 	procBindEnvAndSetDefault(cfg, "process_config.foo.bar", "asdf")
 
 	envs := map[string]struct{}{}
@@ -520,7 +520,7 @@ func TestProcBindEnvAndSetDefault(t *testing.T) {
 }
 
 func TestProcBindEnv(t *testing.T) {
-	cfg := setupConf()
+	cfg := SetupConf()
 	procBindEnv(cfg, "process_config.foo.bar")
 
 	envs := map[string]struct{}{}
@@ -565,7 +565,7 @@ func TestProcConfigEnabledTransform(t *testing.T) {
 		},
 	} {
 		t.Run("process_config.enabled="+tc.procConfigEnabled, func(t *testing.T) {
-			cfg := setupConf()
+			cfg := SetupConf()
 			cfg.Set("process_config.enabled", tc.procConfigEnabled)
 			loadProcessTransforms(cfg)
